refactor: export inference utils
