+++ net/__init__.py
@@ -1,0 +1,3 @@
+# re-export the classes and functions from the other modules
+from .models.SUM import *
+from .inference_utils import *
